==> yelp-text-mining/03-process-counts.ipynb <==
# Text Mining

## Load Packages

# Import packages to use later
import pandas as pd
import nltk
import time
import json

## Process Data

metros = ['Indianapolis', 'Philadelphia', 'Tucson', 'Tampa', 'Nashville']
v2 = True

metro = 'Nashville'
if v2:
    #relationship_df_raw = pd.read_csv("output/" + metro + "_counts_v2.csv")
    relationship_df_raw = pd.read_csv("output_urbcomp/" + metro + "_counts_v2.csv")
else:
    #relationship_df_raw = pd.read_csv("output/" + metro + "_counts.csv")
    relationship_df_raw = pd.read_csv("output_urbcomp/" + metro + "_counts.csv")

relationship_df_raw.head()

# Put relationships into bins ("romantic", "family", etc.) for aggregation
with open("relationships/Relationships_ATUS_custom_v2_binned_update.txt") as f:
    relationships_binned_raw = [line.strip().lower() for line in f.readlines()]

relationships_binned = dict()
for line in relationships_binned_raw:
    rel_type, rel_list = line.split(":")
    relationships_binned[rel_type] = rel_list.split(",")
print(relationships_binned)

# Compute counts for relationship category bins
relationship_df = relationship_df_raw.copy()

for rel_type, rel_list in relationships_binned.items():
    relationship_df[rel_type] = relationship_df[rel_list].sum(axis=1)

relationship_df.head(3)

# Compute normalized counts
relationship_df_norm = relationship_df.copy()
relationship_df_norm = relationship_df_norm[relationship_df_norm["num_reviews"] >= 30]
relationship_df_norm[relationship_df_norm.columns[2:]] = relationship_df_norm[relationship_df_norm.columns[2:]].div(relationship_df_norm.num_reviews, axis=0) * 1000

relationship_df_norm.head(3)

# Compute percentages for each relationship category bin
for relationship in relationships_binned.keys():
    newcol = relationship + "_pct"
    relationship_df[newcol] = relationship_df[relationship].div(relationship_df["num_relationship_words"])

relationship_df.head(3)

### Businesses

# Get top-level business categories
with open("categories.json") as json_file:
    categories = json.load(json_file)

# Dictionary of category keys to the full strings that appear in the dataset
category_dict = {category["alias"]:category["title"] for category in categories}
category_dict_reverse = {category["title"]:category["alias"] for category in categories}

# Categories that we are interested in
#categories_of_interest = ['active', 'arts', 'beautysvc', 'food', 'hotelstravel', 'localflavor', 'nightlife', 'restaurants', 'shopping']
categories_of_interest_raw = [category["alias"] for category in categories if category["parents"] == []]
categories_of_interest = [category_dict[cat] for cat in categories_of_interest_raw]

print(categories_of_interest)
print(categories_of_interest_raw)

# Create indicator columns for each business category (a business may belong to multiple categories)
#businesses = pd.read_csv("yelp_academic_dataset_business_clean.csv")
businesses = pd.read_csv("yelp_academic_dataset_business.csv")
businesses = businesses.rename(columns={"business": "business_id"})

for category_str in categories_of_interest:
    colname = "is_" + category_dict_reverse[category_str]
    businesses[colname] = businesses["categories"].str.contains(category_str).fillna(False)
businesses.head()

## Final Merged Counts Dataframes

relationship_df_final = pd.merge(relationship_df, businesses, how="inner", on="business_id")
relationship_df_final.head(3)

relationship_df_norm_final = pd.merge(relationship_df_norm, businesses, how="inner", on="business_id")
relationship_df_norm_final.head(3)

# Write final CSV files
if v2:
    #relationship_df_final.to_csv("output/" + metro + "_counts_final_v2.csv", index=False)
    #relationship_df_norm_final.to_csv("output/" + metro + "_counts_norm_final_v2.csv", index=False)
    relationship_df_final.to_csv("output_urbcomp/" + metro + "_counts_final_v2.csv", index=False)
    relationship_df_norm_final.to_csv("output_urbcomp/" + metro + "_counts_norm_final_v2.csv", index=False)
else:
    #relationship_df_final.to_csv("output/" + metro + "_counts_final.csv", index=False)
    #relationship_df_norm_final.to_csv("output/" + metro + "_counts_norm_final.csv", index=False)
    relationship_df_final.to_csv("output_urbcomp/" + metro + "_counts_final.csv", index=False)
    relationship_df_norm_final.to_csv("output_urbcomp/" + metro + "_counts_norm_final.csv", index=False)
