==> yelp-text-mining/04-output-files.ipynb <==
import pandas as pd
import matplotlib.pyplot as plt
import json
import numpy as np

## Load Business Category Data

# Get top-level business categories
with open("categories.json") as json_file:
    categories = json.load(json_file)

# Dictionary of category keys to the full strings that appear in the dataset
category_dict = {category["alias"]:category["title"] for category in categories}
category_dict_reverse = {category["title"]:category["alias"] for category in categories}

# Categories that we are interested in
#categories_of_interest = ['active', 'arts', 'beautysvc', 'food', 'hotelstravel', 'nightlife', 'restaurants', 'shopping']
categories_of_interest_raw = [category["alias"] for category in categories if category["parents"] == []]
categories_of_interest = [category_dict[cat] for cat in categories_of_interest_raw]

# print(categories_of_interest)
# print(categories_of_interest_raw)

## Business Data

### Load Dataset

businesses = pd.read_csv("yelp_academic_dataset_business.csv")

businesses.head(3)

### Process Categories

# Create indicator columns for each business category (a business may belong to multiple categories)
businesses_categories = businesses.copy()

for category_str in categories_of_interest:
    colname = "is_" + category_dict_reverse[category_str]
    businesses_categories[colname] = businesses_categories["categories"].str.contains(category_str).fillna(False)
# businesses_categories.head(3)

## Counts Data (all metro areas combined) and num_businesses

### counts_df

metros = ['Indianapolis', 'Philadelphia', 'Tucson', 'Tampa', 'Nashville']
v2 = True

counts_df = pd.DataFrame()
# counts_df_norm = pd.DataFrame()

for metro in metros:
    if v2:
        suffix = "_v2.csv"
    else:
        suffix = ".csv"
        
    #counts_df_metro = pd.read_csv("output/" + metro + "_counts_final" + suffix)
    counts_df_metro = pd.read_csv("output_urbcomp/" + metro + "_counts_final" + suffix)
    counts_df_metro["metroarea"] = metro
    counts_df = counts_df.append(counts_df_metro)
    
#     counts_df_norm_metro = pd.read_csv("output/" + metro + "_counts_norm_final" + suffix)
#     counts_df_norm_metro["metroarea"] = metro
#     counts_df_norm = counts_df_norm.append(counts_df_norm_metro)

# counts_df.head()
#counts_df.to_csv("test/all_counts_final_v2.csv", index=False)
counts_df.to_csv("final_output_urbcomp/all_counts_final_v2.csv", index=False)
# counts_df_norm.to_csv("test/all_counts_norm_final_v2.csv", index=False)

#### num_businesses

categories_of_interest_small = ['active', 'arts', 'beautysvc', 'food', 'hotelstravel', 'nightlife', 'restaurants', 'shopping']
counts_df_mod_noprofs = counts_df.copy()
for col in ["is_"+col for col in categories_of_interest_small if col != "restaurants"]:
    counts_df_mod_noprofs[col] = np.logical_and(counts_df_mod_noprofs[col] == True, counts_df_mod_noprofs["is_restaurants"] == False)
cols = ["is_"+col for col in categories_of_interest_small]
#len(counts_df_mod[counts_df_mod[cols].sum(axis=1) > 1]) / len(counts_df_mod)

df_rows = []
# categories_of_interest_small = ['active', 'arts', 'beautysvc', 'food', 'hotelstravel', 'nightlife', 'restaurants', 'shopping']

cols = [counts_df.columns[0]] + list(counts_df.columns[3:45])
# print(cols)

for category in categories_of_interest_small:
    df_row = [category]
    q = "is_" + category
    counts_category = counts_df_mod_noprofs.query(q)
    num_bus_in_cat = len(counts_category)
    counts_df_temp = counts_category[cols[1:]]
    #print(np.count_nonzero(counts_df_temp))
    df_row.append(num_bus_in_cat)
    df_row.extend(np.count_nonzero(counts_df_temp, axis=0))
    df_rows.append(df_row)
    
num_businesses_df = pd.DataFrame(df_rows,columns = ["category", "num_businesses"] + cols[1:])
    
# num_businesses_df

#num_businesses_df.to_csv("test/num_businesses.csv", index=False)
num_businesses_df.to_csv("final_output_urbcomp/num_businesses.csv", index=False)

### counts_df_norm and table

#### Initial Data Prep

metros = ['Indianapolis', 'Philadelphia', 'Tucson', 'Tampa', 'Nashville']
v2 = True

# counts_df = pd.DataFrame()
counts_df_norm = pd.DataFrame()

for metro in metros:
    if v2:
        suffix = "_v2.csv"
    else:
        suffix = ".csv"
        
#     counts_df_metro = pd.read_csv("output/" + metro + "_counts_final" + suffix)
#     counts_df_metro["metroarea"] = metro
#     counts_df = counts_df.append(counts_df_metro)
    
    #counts_df_norm_metro = pd.read_csv("output/" + metro + "_counts_norm_final" + suffix)
    counts_df_norm_metro = pd.read_csv("output_urbcomp/" + metro + "_counts_norm_final" + suffix)
    counts_df_norm_metro["metroarea"] = metro
    counts_df_norm = counts_df_norm.append(counts_df_norm_metro)

# counts_df.head()
# counts_df.to_csv("test/all_counts_final_v2.csv", index=False)
# counts_df_norm.to_csv("test/all_counts_norm_final_v2.csv", index=False)

#### Table

metros = ['Indianapolis', 'Philadelphia', 'Tucson', 'Tampa', 'Nashville']
metros = sorted(metros)  # , reverse=True)
#metros.append("Total")
#metros.reverse()

# print(metros)

cols = list(counts_df_norm.columns[3:41])
# cols.remove("professor")
# len(cols)#.remove('professor')
# print(cols)

# counts_df_norm = pd.read_csv("output/" + "all" + "_counts_norm_final_v2.csv")
df = pd.DataFrame(counts_df_norm[cols].mean().sort_values(ascending=False))
#df

for metro in metros:
    #counts_df_norm_metro = pd.read_csv("output/" + metro + "_counts_norm_final_v2.csv")
    counts_df_norm_metro = pd.read_csv("output_urbcomp/" + metro + "_counts_norm_final_v2.csv")
    df[metro] = pd.DataFrame(counts_df_norm_metro[cols].mean().sort_values(ascending=False))
df = df.reset_index()
#df
df = df.rename(columns={"index": "Word", 0: "Total"})
df

df.to_latex(buf="output_urbcomp/table.txt", float_format="%.2f", index=False)
#df.to_latex(buf="output/table.txt", float_format="%.2f", index=False)

#### counts_df_norm file

categories_of_interest = ['active', 'arts', 'beautysvc', 'food', 'hotelstravel', 'nightlife', 'restaurants', 'shopping']
q = " or ".join(["is_"+cat for cat in categories_of_interest])
# "(" + q + ") and is_health == False and is_homeservices == False"

final_table = counts_df_norm.query(q)

for col in ["is_"+col for col in categories_of_interest if col != "restaurants"]:
    final_table[col] = np.logical_and(final_table[col] == True, final_table["is_restaurants"] == False)

#final_table.to_csv("test/all_counts_norm_final_v2.csv", index=False)
final_table.to_csv("final_output_urbcomp/all_counts_norm_final_v2.csv", index=False)

#### counts_df_norm_formaps file

#counts_df_norm.head()

cols_of_interest = ['business_id', 'num_reviews', 'num_relationship_words',
                    'family', 'romantic', 'friendship', 'professional',
                    'name', 'categories', 'stars', 'review_count',
                    'address', 'city', 'state', 'longitude', 'latitude', 'metroarea']

cols_of_interest.extend(["is_"+cat for cat in categories_of_interest])
# print(cols_of_interest)

import scipy.stats as stats
counts_df_norm_fewcols = final_table[cols_of_interest]


# Calculate p-values and confidence levels for the z-scores
for col in ['family', 'romantic', 'friendship', 'professional']:
    zscore_col = f"{col}_zscore"
    counts_df_norm_fewcols[zscore_col] = (
        counts_df_norm_fewcols[col] - counts_df_norm_fewcols[col].mean()) / counts_df_norm_fewcols[col].std()
    counts_df_norm_fewcols[f"{col}_pvalue"] = stats.norm.sf(
        abs(counts_df_norm_fewcols[zscore_col])) * 2
    counts_df_norm_fewcols[f"{col}_confidence"] = (
        1 - counts_df_norm_fewcols[f"{col}_pvalue"]) * 100

# Print the confidence levels
# for col in ['family', 'romantic', 'friendship', 'professional']:
#     print(
#         f"{col} confidence level: {counts_df_norm_fewcols[f'{col}_confidence'].mean():.2f}%")
    
print(counts_df_norm_fewcols.head(3))

# businesses_renamed = businesses.rename(columns={"business": "business_id"})
# counts_df_norm_merged = pd.merge(counts_df_norm_fewcols, businesses_renamed, how="inner", on="business_id")
#counts_df_norm_fewcols.to_csv("test/all_counts_norm_fewcols_formaps.csv", index=False)
counts_df_norm_fewcols.to_csv("final_output_urbcomp/all_counts_norm_fewcols_formaps.csv", index=False)
